Clean up imports and add unit tests

+++ sleep/src/main/java/com/noom/interview/fullstack/sleep/service/UserService.java
@@ -1,13 +1,12 @@
 package com.noom.interview.fullstack.sleep.service;
 
+import com.noom.interview.fullstack.sleep.dto.UserDTO;
 import com.noom.interview.fullstack.sleep.entity.User;
 import com.noom.interview.fullstack.sleep.exception.UserNotFoundException;
+import com.noom.interview.fullstack.sleep.mapper.UserMapper;
 import com.noom.interview.fullstack.sleep.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.EntityManager;
 
 @Service
 public class UserService {
@@ -17,22 +16,27 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public User createUser(User user) {
-        return userRepository.save(user);
+    public UserDTO createUser(UserDTO userDTO) {
+        return UserMapper.toDto(userRepository.save(UserMapper.toEntity(userDTO)));
     }
 
-    public User getUserById(Long id){
-        return userRepository.findById(id).orElseThrow(UserNotFoundException::new);
+    public UserDTO getUserById(Long id){
+        return UserMapper.toDto(userRepository.findById(id).orElseThrow(UserNotFoundException::new));
     }
 
-    public User updateUserById(Long id, User user) {
+    public UserDTO updateUserById(Long id, UserDTO user) {
         User existingUser = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
         existingUser.setName(user.getName());
-        return userRepository.save(existingUser);
+        return UserMapper.toDto(userRepository.save(existingUser));
     }
 
     public void deleteUserById(Long id){
-        userRepository.deleteById(id);
+        if(userRepository.existsById(id)){
+            userRepository.deleteById(id);
+        }
+        else{
+            throw new UserNotFoundException();
+        }
     }
 
 }
